TianHaoR {TianHaoR}	R Documentation

Hello,TianHao is here!

Description

Some testing methods of multivariate statistical analysis are given,A method for calculating principal component score in principal component analysis and factor score in factor analysis

Usage


hao.test()
tang.test()
lv.test()
yang.test()
xia.test()
pcascore()
fcascore()

Examples


hao.test(a,miu,cov,alpha)
tang.test(a,miu)
lv.test(x1,x2,alpha)
yang.test(x1,x2,alpha)
xia.test(a,sigma0,alpha)
pcascore(a)
fcascore(a)

[Package TianHaoR version 0.1.0 Index]